+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
 * Issue #43 https://github.com/hapifhir/org.hl7.fhir.validator-wrapper/issues/43
 * Added Terminology API and mockked for tests
 * Updated toggle buttons https://github.com/hapifhir/org.hl7.fhir.validator-wrapper/issues/38
+* Added search functionality for IG Dropdown https://github.com/hapifhir/org.hl7.fhir.validator-wrapper/issues/41